Refactor rendering and game systems
- Refactor renderer into tile renderer and sprite renderer.
- Pull out renderers and connection into a game class.
- Remove renderer from connection.
- Handle all packets pending on connection tick.
- Fix velocity in player data when reading position packets.
- Clean up main file and update window in main.
- Send disconnect packet when closing the window.
- Track weapon's received to obey S2C weapon packet loss check.
- Move MD5 and checksum into net folder.

+++ src/net/PacketSequencer.cpp
@@ -44,7 +44,7 @@ void PacketSequencer::Tick(Connection& connection) {
   u32 current_tick = GetCurrentTick();
 
   // Resend timed out messages from reliable_sent
-  for (int i = 0; i < reliable_sent_count; ++i) {
+  for (size_t i = 0; i < reliable_sent_count; ++i) {
     ReliableMessage* mesg = reliable_sent + i;
 
     if (TICK_DIFF(current_tick, mesg->timestamp) >= kResendDelay) {
